Improve handling of invalid values and various other changes
Throw a FormatException for unsupported bit depths in the
ExpandPackedPalette method.
Throw a FormatException for invalid document dimensions.
Skip layers with invalid dimensions.

Rename the bpp variable and consolidate the bit depth if statements
into a switch statement in the Load method.
Consolidate the transparency related ternary operators into
an if statement in the Save method.