Added save constructor overloads and skip trailing data
Added save constructor overloads to the LayerInfoChunk and
LayerBitmapChunk classes.
Added a HeaderSize constant and save constructor overload
to the LayerBitmapInfo class.
Skip the expansion field when reading the LayerBitmapInfo chunk.

Removed the fileMajorVersion field from the LayerInfoChunk class
and added a save constructor overload.
Added a majorVersion parameter to the LayerBitmapInfo Save method.