Allow empty layers to be saved and skip the unknown layer block
Changed the HeaderSize constant to internal in the
LayerBitmapInfoChunk class.
Replaced the GeneralImageAttributes parameter with a
PSPCompression parameter and check the block signature and type 
in the LayerBitmapInfoChunk constructors.

Added Version5ChunkLength and Version6ChunkLength constants and set
LayerInfoChunk class.
Set the layer type in the LayerInfoChunk(PaintDotNet.Layer) overload.

Updated the LayerBlock constructor to use the new
LayerBitmapInfoChunk constructor.
Save the the LayerBitmapInfo for empty layers and updated the
chunk length field comments in the LayerBlock Save method.

Attempt to skip the unknown block type placed before the start of the
LayerBitmapInfoChunk in Paint Shop Pro X2 and later files.